Always move in journal when seeking to tail
According to [this Github issue][1] on the systemd-journal gem, you must
make one movement in the journal before methods like current_entry and
cursor are valid. Although neither are called in this plugin, the watch
method also behaves weirdly, in my testing it always starts from the
beginning of the journal even when given `seekto => "tail"`.

This appears to be a limitation of the journal C library itself, the
[documentation for which][2] states:

> Note that these functions will not work before sd_journal_next(3) (or
> related call) has been called at least once, in order to position the
> read pointer at a valid entry.

I tested this with my local CentOS 7 journal and verified that seeking
to the tail and then calling move_previous does not actually seek to
repeat the previous event.

[1]: https://github.com/ledbettj/systemd-journal/issues/55
[2]: http://www.freedesktop.org/software/systemd/man/sd_journal_get_data.html